# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ft_transcendence API
  description: >
    API REST du backend ft_transcendence (Fastify v5 + TypeScript).
    Auth JWT (access token Bearer 15 min) + refresh token en cookie httpOnly
    (SameSite=Strict). Le cookie est posé Path=/auth par le backend et réécrit en
    Path=/api/auth par le proxy Vite pour le navigateur (invariant I4). Chat temps réel
    hors-OpenAPI via WebSocket `/ws/chat` (documenté en fin de fichier pour mémoire).
  version: 0.9.0
  contact:
    name: ft_transcendence team
servers:
  # Origine applicative PRINCIPALE (navigateur) : tout passe par le proxy Vite same-origin.
  # Les chemins ci-dessous sont relatifs à cette base — /auth/login s'appelle donc
  # https://localhost:5173/api/auth/login côté navigateur.
  - url: https://localhost:5173/api
    description: App proxifiée (origine unique navigateur, I4)
  # Backend en direct : diagnostic, tests e2e Python et curl uniquement (hors navigateur app).
  - url: https://localhost:3000
    description: Backend direct (diagnostic / tests, HTTPS certificat auto-signé)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token signé (payload `{ sub }`). Les tempTokens 2FA
        (`{ sub, pending: 'totp' }`) sont refusés sur les routes protégées.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Unauthorized
      required: [error]
    ValidationError:
      type: object
      description: Erreur de validation Zod (issues brutes).
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: too_small
              path:
                type: array
                items:
                  type: string
              message:
                type: string
                example: au moins 8 caractères
      required: [errors]
...
    Game:
      type: object
      description: Un jeu supporté par la plateforme.
      properties:
        id:
          type: string
          description: Slug du jeu (PK).
          example: cs2
        name:
          type: string
          example: Counter-Strike 2
        requiredProvider:
          type: string
          enum: [riot, steam, epic, chess_com]
          description: Provider d'identifiant in-game requis pour s'inscrire à un match de ce jeu.
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, requiredProvider, isActive, createdAt]
    Ladder:
      type: object
      description: Une compétition (jeu × format) avec son ELO indépendant.
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          example: cs2
        format:
          type: string
          enum: ['1v1', '2v2', '3v3', '5v5']
        name:
          type: string
          example: Counter-Strike 2 5v5
        lockoutMinutes:
          type: integer
          description: >
            Durée pendant laquelle la team/user est verrouillée après le début
            d'un match. Aussi le délai minimum avant soumission de score.
          example: 60
        createdAt:
          type: string
          format: date-time
      required: [id, gameId, format, name, lockoutMinutes, createdAt]

    # ---------- External accounts (in-game) ----------
...
  responses:
...
    InternalError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /ladders:
    get:
      tags: [Ladders]
      summary: Liste les ladders
      description: >
        Renvoie tous les ladders triés par nom. Filtre optionnel par
        `gameId` (slug du jeu). Un `gameId` inconnu renvoie une liste vide
        (pas d'erreur).
      parameters:
        - name: gameId
          in: query
          required: false
          schema:
            type: string
            example: cs2
          description: Filtre les ladders pour un jeu donné.
      responses:
        '200':
          description: Liste des ladders
          content:
            application/json:
              schema:
                type: object
                properties:
                  ladders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ladder'
                required: [ladders]
        '500':
          $ref: '#/components/responses/InternalError'
  /ladders/{id}:
    get:
      tags: [Ladders]
      summary: Détails d'un ladder (avec son jeu joint)
      description: >
        Lookup par uuid du ladder. Renvoie le ladder + le jeu parent
        (JOIN pour éviter au frontend un 2e appel sur `GET /games/{id}`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Le ladder demandé et son jeu parent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ladder:
                    $ref: '#/components/schemas/Ladder'
                  game:
                    $ref: '#/components/schemas/Game'
                required: [ladder, game]
        '400':
          description: Param :id non-uuid (ValidationError)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Ladder inconnu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /ladders/{id}/rankings:
    get:
      tags: [Ladders]
      summary: Classement (leaderboard) d'un ladder
      description: >
        Renvoie le classement du ladder, trié par ELO décroissant. Chaque
        entrée porte un `competitor` qui est soit un joueur (`type: user`),
        soit une équipe (`type: team`) selon le ladder. Un ladder existant
        sans classement renvoie une liste vide (pas d'erreur).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Classement trié par ELO décroissant (peut être vide)
          content:
            application/json:
              schema:
                type: object
                properties:
                  rankings:
                    type: array
                    items:
                      type: object
                      properties:
                        rank:
                          type: integer
                          example: 1
                        elo:
                          type: integer
                          example: 1580
                        wins:
                          type: integer
                        losses:
                          type: integer
                        competitor:
                          oneOf:
                            - type: object
                              properties:
                                type:
                                  type: string
                                  enum: [user]
                                pseudo:
                                  type: string
                                displayName:
                                  type: string
                                  nullable: true
                                avatarUrl:
                                  type: string
                                  nullable: true
                              required: [type, pseudo]
                            - type: object
                              properties:
                                type:
                                  type: string
                                  enum: [team]
                                name:
                                  type: string
                                logoUrl:
                                  type: string
                                  nullable: true
                              required: [type, name]
                      required: [rank, elo, wins, losses, competitor]
                required: [rankings]
        '400':
          description: Param :id non-uuid (ValidationError)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Ladder inconnu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  # =====================================================================
  # TEAMS
  # =====================================================================
...